Refactor PlayerCards State for Strategy Training
Restructured the PlayerCards state in the StrategyTraining module to
utilize an array of objects that represent hands the player currently
holds on the game board. This refactor enhances the representation of
player hands and provides flexibility for future features, such as
hand splitting. In the event that a player decides to split a hand,
an additional object can now be seamlessly added to the PlayerCards
array.