Enhancements and Bug Fixes in Blackjack Gameplay
In this commit:

Implemented dealer check and dealer turn states within the Finite
State Machine (FSM).
Introduced a hand total for the dealer's hand to aid in debugging.
Resolved issues related to the correct functioning of the Insurance
feature.
Addressed certain bugs; however, further work is needed for scenarios
where both the player and the dealer have a hand value of 21.